+++ modules/app/src/main/scala/cards/nine/app/ui/collections/actions/recommendations/RecommendationsFragment.scala
@@ -11,7 +11,7 @@ import cards.nine.commons.services.TaskService
 import cards.nine.commons.services.TaskService.{TaskService, _}
 import cards.nine.models.types.NineCardsCategory
 import cards.nine.process.recommendations.RecommendedAppsConfigurationException
-import cards.nine.process.recommendations.models.RecommendedApp
+import cards.nine.models.RecommendedApp
 import com.fortysevendeg.ninecardslauncher.R
 
 class RecommendationsFragment(implicit groupCollectionsJobs: GroupCollectionsJobs, singleCollectionJobs: Option[SingleCollectionJobs])