+++ modules/app/src/main/scala/cards/nine/app/ui/collections/dialog/publishcollection/PublishCollectionJobs.scala
@@ -1,15 +1,12 @@
 package cards.nine.app.ui.collections.dialog.publishcollection
 
-import cards.nine.models.types.NineCardsCategory
 import cats.implicits._
-import com.fortysevendeg.macroid.extras.ResourcesExtras._
 import cards.nine.app.ui.commons.{JobException, Jobs}
 import cards.nine.commons.services.TaskService
 import cards.nine.commons.services.TaskService._
 import cards.nine.models.Collection
 import cards.nine.models.types.NineCardsCategory
 import cards.nine.process.sharedcollections.SharedCollectionsException
-import cards.nine.process.sharedcollections.models.CreateSharedCollection
 import com.fortysevendeg.macroid.extras.ResourcesExtras._
 import com.fortysevendeg.ninecardslauncher.R
 import macroid.ActivityContextWrapper
@@ -44,14 +41,13 @@ class PublishCollectionJobs(actions: PublishCollectionActions)(implicit val cont
       for {
         user <- di.userProcess.getUser
         collection <- getCollection
-        sharedCollection = CreateSharedCollection(
+        sharedCollectionId <- di.sharedCollectionsProcess.createSharedCollection(
           author = user.userProfile.name getOrElse (user.email getOrElse resGetString(R.string.defaultUser)),
           name = name,
           packages = collection.cards flatMap (_.packageName),
           category = category,
           icon = collection.icon,
           community = false)
-        sharedCollectionId <- di.sharedCollectionsProcess.createSharedCollection(sharedCollection)
         _ <- di.collectionProcess.updateSharedCollection(collection.id, sharedCollectionId)
       } yield sharedCollectionId