+++ modules/app/src/main/scala/cards/nine/app/ui/launcher/exceptions/LauncherExceptions.scala
@@ -2,7 +2,6 @@ package cards.nine.app.ui.launcher.exceptions
 
 import cards.nine.commons.services.TaskService.NineCardException
 
-
 case class SpaceException(message: String, cause : Option[Throwable] = None)
   extends RuntimeException(message)
   with NineCardException {
@@ -11,4 +10,28 @@ case class SpaceException(message: String, cause : Option[Throwable] = None)
 
 trait ImplicitsSpaceException {
   implicit def momentException = (t: Throwable) => SpaceException(t.getMessage, Option(t))
+}
+
+case class LoadDataException(message: String, cause: Option[Throwable] = None, recoverable: Boolean = false)
+  extends RuntimeException(message)
+  with NineCardException {
+
+  cause foreach initCause
+
+}
+
+trait ImplicitsLoadDataException {
+  implicit def loadDataExceptionConverter = (t: Throwable) => LoadDataException(t.getMessage, Option(t))
+}
+
+case class ChangeMomentException(message: String, cause: Option[Throwable] = None, recoverable: Boolean = false)
+  extends RuntimeException(message)
+  with NineCardException {
+
+  cause foreach initCause
+
+}
+
+trait ImplicitsChangeMomentException {
+  implicit def changeMomentExceptionConverter = (t: Throwable) => ChangeMomentException(t.getMessage, Option(t))
 }