+++ modules/models/src/main/scala/cards/nine/models/Awareness.scala
@@ -1,5 +1,7 @@
 package cards.nine.models
 
+import cards.nine.models.types.{ConditionWeather, KindActivity}
+
 case class ProbablyActivity(activityType: KindActivity)
 
 case class Headphones(connected: Boolean)
@@ -28,43 +30,3 @@ case class Location(
   countryCode: Option[String],
   countryName: Option[String],
   addressLines: Seq[String])
-
-sealed trait KindActivity
-
-case object InVehicleActivity extends KindActivity
-
-case object OnBicycleActivity extends KindActivity
-
-case object OnFootActivity extends KindActivity
-
-case object RunningActivity extends KindActivity
-
-case object StillActivity extends KindActivity
-
-case object TiltingActivity extends KindActivity
-
-case object WalkingActivity extends KindActivity
-
-case object UnknownActivity extends KindActivity
-
-sealed trait ConditionWeather
-
-case object ClearCondition extends ConditionWeather
-
-case object CloudyCondition extends ConditionWeather
-
-case object FoggyCondition extends ConditionWeather
-
-case object HazyCondition extends ConditionWeather
-
-case object IcyCondition extends ConditionWeather
-
-case object RainyCondition extends ConditionWeather
-
-case object SnowyCondition extends ConditionWeather
-
-case object StormyCondition extends ConditionWeather
-
-case object WindyCondition extends ConditionWeather
-
-case object UnknownCondition extends ConditionWeather