+++ modules/models/src/main/scala/cards/nine/models/types/Permission.scala
@@ -1,4 +1,4 @@
-package cards.nine.models
+package cards.nine.models.types
 
 sealed trait PermissionStatus