+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
     * Change backend URL at runtime.
     * Enable/disable Stetho.
 * Refactorized all persistence models.
+* Refactorized all api services models.
 
 ## 2016/10/14 - Version 2.0-alpha